third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-summary-show-next-expected.txt

Tests Summary view of detailed heap snapshots. Repeated clicks on "Show Next" button must show all nodes.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowNext

Profiler was disabled.
